Eligibility Determination Service modified

+++ src/main/java/com/jrtp/ies/dc/model/CasePlan.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class CasePlan {
 
-	private short caseId;
+	private int caseId;
 	
 	private String firstName;
